requirements.txt
PyAudio==0.2.14
pyinstaller==6.7.0
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
requests==2.31.0
speechrecognition==3.10.4